move logical to PDFManager
