+++ src/components/specific/models/models-manager/pdf-manager/PDFManager.vue
@@ -1,5 +1,10 @@
 <template>
-  <GenericModelsManager :project="project" :tabs="tabs">
+  <GenericModelsManager
+    :project="project"
+    :tabs="tabs"
+    @edit="editModel"
+    @get-current-tab="currentTab = $event"
+  >
     <template #tablePlaceholder v-if="models.length === 0">
       <div class="pdf-manager__placeholder">
         <div
@@ -14,12 +19,46 @@
       </div>
     </template>
   </GenericModelsManager>
+  <template v-if="currentTab && currentTab.id === 'metaBuildings'">
+    <BIMDataButton
+      class="pdf-manager__building-maker-btn"
+      width="100px"
+      color="primary"
+      outline
+      radius
+      icon
+      @click="openBM"
+    >
+      <BIMDataIcon name="building" size="xxxs" margin="0 12px 0 0" />
+      <span> {{ $t("ModelsManager.buildingMaker.create") }} </span>
+    </BIMDataButton>
+
+    <!-- <transition name="slide-fade-right"> -->
+    <div v-if="isBMOpen" class="pdf-manager__building-maker">
+      <div class="pdf-manager__building-maker__header">
+        <BIMDataButton ghost rounded icon @click="closeBM">
+          <BIMDataIcon name="close" size="xxs" fill color="granite-light" />
+        </BIMDataButton>
+      </div>
+      <div class="pdf-manager__building-maker__content">
+        <BuildingMaker
+          :apiClient="apiClient"
+          :space="project.cloud"
+          :project="project"
+          :model="currentModel"
+        />
+      </div>
+    </div>
+    <!-- </transition> -->
+  </template>
 </template>
 
 <script>
 import { computed, ref, watch } from "vue";
 import { useI18n } from "vue-i18n";
 import { MODEL_TYPE } from "../../../../../config/models.js";
+import { useToggle } from "../../../../../composables/toggle";
+import apiClient from "../../../../../services/api-client.js";
 import { segregateBySource } from "../../../../../utils/models.js";
 // Components
 import GenericModelsManager from "../generic-models-manager/GenericModelsManager.vue";
@@ -43,6 +82,17 @@ export default {
   setup(props) {
     const { locale, fallbackLocale } = useI18n();
     const tabs = ref(tabsDef);
+    const currentTab = ref(null);
+
+    const { isOpen: isBMOpen, close: closeBM, open: openBM } = useToggle();
+
+    const currentModel = ref(null);
+
+    const editModel = async model => {
+      currentModel.value = model;
+      openBM();
+      setTimeout(() => (currentModel.value = null), 100);
+    };
 
     watch(
       () => props.models,
@@ -72,8 +122,18 @@ export default {
     );
 
     return {
+      // References
+      currentModel,
+      isBMOpen,
       placeholderBackground,
-      tabs
+      tabs,
+      apiClient,
+      currentTab,
+      // Methods
+      openBM,
+      closeBM,
+      editModel,
+      console
     };
   }
 };
